requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
dj-database-url==1.1.0
Django==5.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-widget-tweaks==1.5.0
freetype-py==2.3.0
fsspec==2025.12.0
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
packaging==25.0
pillow==10.2.0
propcache==0.4.1
psycopg2-binary==2.9.9
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.0
pypng==0.20220715.0
pyroaring==1.0.3
python-dateutil==2.8.2
python-decouple==3.8
qrcode==7.4.2
reportlab==4.0.0
requests==2.32.5
rich==14.2.0
rlPyCairo==0.4.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
StrEnum==0.4.15
strictyaml==1.7.3
# Supabase-related packages removed — project uses Django storage and Postgres now
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
websockets==15.0.1
whitenoise==6.6.0
yarl==1.22.0

# Ensure python-dotenv is available on the server so .env files can be loaded
python-dotenv==1.0.0
